# deeptector-backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Deeptector API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/sessions": {
      "post": {
        "summary": "Create Session",
        "operationId": "create_session_api_v1_sessions_post",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses": {
      "post": {
        "summary": "Queue Analysis",
        "operationId": "queue_analysis_api_v1_analyses_post",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_queue_analysis_api_v1_analyses_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisDetail"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ]
      },
      "get": {
        "summary": "List Analyses",
        "operationId": "list_analyses_api_v1_analyses_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "status": {
            "type": "string",
            "enum": [
              "QUEUED",
              "PROCESSING",
              "COMPLETED",
              "FAILED"
            ],
            "title": "Status"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactResponse"
            },
            "type": "array",
            "title": "Artifacts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "status",
          "stage",
          "created_at",
          "updated_at",
          "expires_at"
        ],
        "title": "AnalysisDetail"
      },
      "AnalysisListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnalysisSummary"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "AnalysisListResponse"
      },
      "AnalysisSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "status": {
            "type": "string",
            "enum": [
              "QUEUED",
              "PROCESSING",
              "COMPLETED",
              "FAILED"
            ],
            "title": "Status"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "status",
          "stage",
          "created_at",
          "updated_at",
          "expires_at"
        ],
        "title": "AnalysisSummary"
      },
...
      "ArtifactResponse": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "mime_type",
          "url"
        ],
        "title": "ArtifactResponse"
      },
      "Body_queue_analysis_api_v1_analyses_post": {
        "properties": {
          "video": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Video",
            "description": "Video to analyze"
          }
        },
        "type": "object",
        "required": [
          "video"
        ],
        "title": "Body_queue_analysis_api_v1_analyses_post"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelResult": {
        "properties": {
          "spatial": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Spatial"
          },
          "temporal": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Temporal"
          },
          "fused": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Fused"
          },
          "n_faces": {
            "type": "integer",
            "maximum": 16.0,
            "minimum": 4.0,
            "title": "N Faces"
          },
          "family": {
            "type": "string",
            "title": "Family"
          },
          "faithfulness": {
            "type": "string",
            "title": "Faithfulness"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "model_version": {
            "type": "string",
            "minLength": 1,
            "title": "Model Version"
          },
          "calibration_version": {
            "type": "string",
            "minLength": 1,
            "title": "Calibration Version"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "is_demo": {
            "type": "boolean",
            "title": "Is Demo"
          },
          "policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskPolicy"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "spatial",
          "temporal",
          "fused",
          "n_faces",
          "family",
          "faithfulness",
          "metadata",
          "model_version",
          "calibration_version",
          "warnings",
          "is_demo"
        ],
        "title": "ModelResult"
      },
      "RiskPolicy": {
        "properties": {
          "verdict": {
            "type": "string",
            "enum": [
              "REAL",
              "FAKE"
            ],
            "title": "Verdict"
          },
          "traffic_light": {
            "type": "string",
            "enum": [
              "RED",
              "YELLOW",
              "GREEN"
            ],
            "title": "Traffic Light"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "primary_signal": {
            "type": "string",
            "enum": [
              "SPATIAL",
              "TEMPORAL",
              "BALANCED"
            ],
            "title": "Primary Signal"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version"
          },
          "signal_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Signal Score"
          },
          "thresholds": {
            "$ref": "#/components/schemas/Thresholds"
          },
          "score_interpretation": {
            "type": "string",
            "title": "Score Interpretation"
          }
        },
        "type": "object",
        "required": [
          "verdict",
          "traffic_light",
          "reason",
          "primary_signal",
          "policy_version",
          "signal_score",
          "thresholds",
          "score_interpretation"
        ],
        "title": "RiskPolicy"
      },
      "SessionResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "token",
          "expires_at"
        ],
        "title": "SessionResponse"
      },
...
      "Thresholds": {
        "properties": {
          "fused_fake": {
            "type": "number",
            "title": "Fused Fake"
          },
          "branch_fake": {
            "type": "number",
            "title": "Branch Fake"
          },
          "branch_caution": {
            "type": "number",
            "title": "Branch Caution"
          }
        },
        "type": "object",
        "required": [
          "fused_fake",
          "branch_fake",
          "branch_caution"
        ],
        "title": "Thresholds"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "SessionBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "POST /api/v1/sessions response token"
      }
    }
  }
}